refactor: switch to more efficient IstioOperator conversion
