+++ istioctl/pkg/install/verify.go
@@ -37,6 +37,7 @@ import (
 	"istio.io/istio/operator/cmd/mesh"
 	"istio.io/istio/operator/pkg/apis/istio/v1alpha1"
 	"istio.io/istio/operator/pkg/controlplane"
+	"istio.io/istio/operator/pkg/object"
 	"istio.io/istio/operator/pkg/translate"
 	"istio.io/istio/pilot/pkg/serviceregistry/kube/controller"
 	"istio.io/istio/pkg/config/schema"
@@ -180,7 +181,7 @@ func verifyPostInstall(enableVerbose bool, istioNamespaceFlag string,
 			// usual conversion not available.  Convert unstructured to string
 			// and ask operator code to unmarshal.
 
-			iop, err := convertToIstioOperator(un.Object)
+			iop, err := object.ConvertUnstructuredToIstioOperator(un.Object)
 			if err != nil {
 				return err
 			}
@@ -416,7 +417,7 @@ func operatorFromCluster(istioNamespaceFlag string, revision string, restClientG
 		return nil, err
 	}
 	for _, un := range ul.Items {
-		iop, err := convertToIstioOperator(un.Object)
+		iop, err := object.ConvertUnstructuredToIstioOperator(un.Object)
 		if err != nil {
 			return nil, err
 		}
@@ -425,15 +426,6 @@ func operatorFromCluster(istioNamespaceFlag string, revision string, restClientG
 		}
 	}
 	return nil, fmt.Errorf("control plane revision %q not found", revision)
-}
-
-func convertToIstioOperator(obj map[string]interface{}) (*v1alpha1.IstioOperator, error) {
-	out := &v1alpha1.IstioOperator{}
-	err := runtime.DefaultUnstructuredConverter.FromUnstructured(obj, out)
-	if err != nil {
-		return nil, err
-	}
-	return out, nil
 }
 
 // Find all IstioOperator in the cluster.
@@ -446,7 +438,7 @@ func allOperatorsInCluster(client dynamic.Interface) ([]*v1alpha1.IstioOperator,
 	}
 	retval := make([]*v1alpha1.IstioOperator, 0)
 	for _, un := range ul.Items {
-		iop, err := convertToIstioOperator(un.Object)
+		iop, err := object.ConvertUnstructuredToIstioOperator(un.Object)
 		if err != nil {
 			return nil, err
 		}
