+++ operator/pkg/validate/common.go
@@ -24,12 +24,10 @@ import (
 
 	"github.com/ghodss/yaml"
 
-	meta_v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
-
 	"istio.io/pkg/log"
 
 	"istio.io/istio/operator/pkg/apis/istio/v1alpha1"
+	"istio.io/istio/operator/pkg/object"
 	"istio.io/istio/operator/pkg/util"
 )
 
@@ -274,19 +272,11 @@ type ValidatorFunc func(path util.Path, i interface{}) util.Errors
 
 // UnmarshalIOP unmarshals a string containing IstioOperator as YAML.
 func UnmarshalIOP(iopYAML string) (*v1alpha1.IstioOperator, error) {
-	// Remove creationDate (util.UnmarshalWithJSONPB fails if present)
 	mapIOP := make(map[string]interface{})
 	if err := yaml.Unmarshal([]byte(iopYAML), &mapIOP); err != nil {
 		return nil, err
 	}
-	un := &unstructured.Unstructured{Object: mapIOP}
-	un.SetCreationTimestamp(meta_v1.Time{}) // UnmarshalIstioOperator chokes on these
-	iopYAML = util.ToYAML(un)
-	iop := &v1alpha1.IstioOperator{}
-	if err := util.UnmarshalWithJSONPB(iopYAML, iop, false); err != nil {
-		return nil, fmt.Errorf("%s:\n\nYAML:\n%s", err, iopYAML)
-	}
-	return iop, nil
+	return object.ConvertUnstructuredToIstioOperator(mapIOP)
 }
 
 // ValidIOP validates the given IstioOperator object.
